Moved snap to anchor functions to its own file
