+++ src/components/nodes/boundaryEvent/boundaryEvent.vue
@@ -11,38 +11,7 @@ import { id as callActivityId } from '@/components/nodes/callActivity';
 import { id as manualTaskId } from '@/components/nodes/manualTask';
 import { id as scriptTaskId } from '@/components/nodes/scriptTask';
 import portsConfig from '@/mixins/portsConfig';
-import { portGroups } from '@/mixins/portsConfig';
-import { g } from 'jointjs';
-
-function getPointFromGroup(model, group) {
-  const { x: shapeX, y: shapeY } = model.position();
-  const { x, y } = Object.values(model.getPortsPositions(group))[0];
-  return new g.Point(shapeX + x, shapeY + y);
-}
-
-function getPortPoints(model) {
-  return portGroups.filter(group => group !== 'absolute').map(group => getPointFromGroup(model, group));
-}
-
-function closestPort(model, anchorReference) {
-  return getPortPoints(model).sort((p1, p2) => {
-    const referencePoint = new g.Point(anchorReference.x, anchorReference.y);
-    return referencePoint.distance(p1) - referencePoint.distance(p2);
-  })[0];
-}
-
-function hasPorts(model) {
-  return Object.values(model.getPortsPositions(portGroups[0])).length > 0;
-}
-
-function snapToAnchor(coords, model) {
-  if (!hasPorts(model)) {
-    const { x, y } = model.position();
-    const { width, height } = model.size();
-    return new g.Point(x - (width / 2), y - (height / 2));
-  }
-  return closestPort(model, coords);
-}
+import { getAnchorCoordinates } from '@/snapToAnchor';
 
 export default {
   props: ['graph', 'node'],
@@ -111,9 +80,9 @@ export default {
       this.shape.attr('label/text', this.node.definition.get('name'));
       this.shape.component = this;
     },
-    updateSnappingPosition(task) {
-      const {x, y} = snapToAnchor(this.shape.position(), task);
-      const {width} = this.shape.size();
+    updateShapePosition(task) {
+      const { x, y } = getAnchorCoordinates(this.shape.position(), task);
+      const { width } = this.shape.size();
       this.shape.position(x - (width / 2), y - (width / 2));
       this.updateCrownPosition();
     },
@@ -127,13 +96,13 @@ export default {
       task.embed(this.shape);
       this.node.definition.set('attachedToRef', task.component.node.definition);
       this.toggleInterruptingStyle(this.node.definition.cancelActivity);
-      this.updateSnappingPosition(task);
+      this.updateShapePosition(task);
 
       this.shape.listenTo(this.paper, 'element:pointerup', cellView => {
         if (cellView.model !== this.shape) {
           return;
         }
-        this.updateSnappingPosition(task);
+        this.updateShapePosition(task);
       });
     },
   },
